+++ mountpoint-s3-client/src/lib.rs
@@ -1,6 +1,6 @@
 mod build_info;
 pub mod checksums;
-mod endpoint;
+mod endpoint_config;
 pub mod failure_client;
 mod imds_crt_client;
 pub mod mock_client;
@@ -8,7 +8,7 @@ mod object_client;
 mod s3_crt_client;
 mod util;
 
-pub use endpoint::{AddressingStyle, Endpoint};
+pub use endpoint_config::{AddressingStyle, EndpointConfig};
 pub use imds_crt_client::ImdsCrtClient;
 pub use object_client::*;
 pub use s3_crt_client::head_bucket::HeadBucketError;
@@ -20,6 +20,6 @@ mod tests {
 
     #[test]
     fn smoke() {
-        let _client = S3CrtClient::new("us-east-1", Default::default()).unwrap();
+        let _client = S3CrtClient::new(Default::default()).unwrap();
     }
 }