+++ mountpoint-s3-client/tests/head_bucket.rs
@@ -3,7 +3,7 @@
 pub mod common;
 
 use common::*;
-use mountpoint_s3_client::{HeadBucketError, ObjectClientError, S3CrtClient};
+use mountpoint_s3_client::{EndpointConfig, HeadBucketError, ObjectClientError, S3ClientConfig, S3CrtClient};
 
 #[tokio::test]
 async fn test_head_bucket_correct_region() {
@@ -15,8 +15,9 @@ async fn test_head_bucket_correct_region() {
 
 #[tokio::test]
 async fn test_head_bucket_wrong_region() {
-    let secondary_region = get_secondary_test_region();
-    let client = S3CrtClient::new(&secondary_region, Default::default()).expect("could not create test client");
+    let endpoint_config = EndpointConfig::new(&get_secondary_test_region());
+    let client =
+        S3CrtClient::new(S3ClientConfig::new().endpoint_config(endpoint_config)).expect("could not create test client");
     let (bucket, _) = get_test_bucket_and_prefix("test_head_bucket_wrong_region");
     let expected_region = get_test_region();