+++ mountpoint-s3-client/tests/put_object.rs
@@ -5,6 +5,7 @@ pub mod common;
 use common::*;
 use futures::{pin_mut, StreamExt};
 use mountpoint_s3_client::checksums::crc32c_to_base64;
+use mountpoint_s3_client::EndpointConfig;
 use mountpoint_s3_client::GetObjectError;
 use mountpoint_s3_client::ObjectClient;
 use mountpoint_s3_client::ObjectClientError;
@@ -186,8 +187,10 @@ async fn test_put_object_abort() {
 async fn test_put_checksums() {
     const PART_SIZE: usize = 5 * 1024 * 1024;
     let (bucket, prefix) = get_test_bucket_and_prefix("test_put_checksums");
-    let client_config = S3ClientConfig::new().part_size(PART_SIZE);
-    let client = S3CrtClient::new(&get_test_region(), client_config).expect("could not create test client");
+    let client_config = S3ClientConfig::new()
+        .part_size(PART_SIZE)
+        .endpoint_config(EndpointConfig::new(&get_test_region()));
+    let client = S3CrtClient::new(client_config).expect("could not create test client");
     let key = format!("{prefix}hello");
 
     let mut rng = rand::thread_rng();
@@ -229,8 +232,10 @@ async fn test_put_checksums() {
 async fn test_put_review(pass_review: bool) {
     const PART_SIZE: usize = 5 * 1024 * 1024;
     let (bucket, prefix) = get_test_bucket_and_prefix("test_put_review");
-    let client_config = S3ClientConfig::new().part_size(PART_SIZE);
-    let client = S3CrtClient::new(&get_test_region(), client_config).expect("could not create test client");
+    let client_config = S3ClientConfig::new()
+        .part_size(PART_SIZE)
+        .endpoint_config(EndpointConfig::new(&get_test_region()));
+    let client = S3CrtClient::new(client_config).expect("could not create test client");
     let key = format!("{prefix}hello");
 
     let mut rng = rand::thread_rng();