+++ mountpoint-s3/examples/fs_benchmark.rs
@@ -2,7 +2,7 @@ use clap::{Arg, ArgAction, Command};
 use fuser::{BackgroundSession, MountOption, Session};
 use mountpoint_s3::fuse::S3FuseFilesystem;
 use mountpoint_s3::S3FilesystemConfig;
-use mountpoint_s3_client::{S3ClientConfig, S3CrtClient};
+use mountpoint_s3_client::{EndpointConfig, S3ClientConfig, S3CrtClient};
 use mountpoint_s3_crt::common::rust_log_adapter::RustLogAdapter;
 use std::{
     fs::File,
@@ -146,11 +146,11 @@ fn mount_file_system(bucket_name: &str, region: &str, throughput_target_gbps: Op
     let temp_dir = tempdir().expect("Should be able to create temp directory");
     let mountpoint = temp_dir.path();
 
-    let mut config = S3ClientConfig::new();
+    let mut config = S3ClientConfig::new().endpoint_config(EndpointConfig::new(region));
     if let Some(throughput_target_gbps) = throughput_target_gbps {
         config = config.throughput_target_gbps(throughput_target_gbps);
     }
-    let client = S3CrtClient::new(region, config).expect("Failed to create S3 client");
+    let client = S3CrtClient::new(config).expect("Failed to create S3 client");
     let runtime = client.event_loop_group();
 
     let mut options = vec![MountOption::RO, MountOption::FSName("mountpoint-s3".to_string())];