+++ mountpoint-s3/tests/cli.rs
@@ -122,25 +122,11 @@ fn print_version_short() -> Result<(), Box<dyn std::error::Error>> {
 }
 
 #[test]
-fn addressing_style_mutually_exclusive() -> Result<(), Box<dyn std::error::Error>> {
-    let mut cmd = Command::cargo_bin("mount-s3")?;
-
-    cmd.arg("test-bucket")
-        .arg("test/dir")
-        .arg("--virtual-addressing")
-        .arg("--path-addressing");
-    let error_message = "the argument '--virtual-addressing' cannot be used with '--path-addressing'";
-    cmd.assert().failure().stderr(predicate::str::contains(error_message));
-
-    Ok(())
-}
-
-#[test]
 fn bucket_name_and_directory_swapped() -> Result<(), Box<dyn std::error::Error>> {
     let mut cmd = Command::cargo_bin("mount-s3")?;
 
     cmd.arg("test/dir").arg("my-bucket-name");
-    let error_message = "bucket names can only contain letters, numbers, . and -";
+    let error_message = "bucket argument should be a valid bucket name(only letters, numbers, . and -) or a valid ARN";
     cmd.assert().failure().stderr(predicate::str::contains(error_message));
 
     Ok(())